fixed a bug with local properties, added some misc things
Local properties (properties stored directly in the custom data part of an object) previously got ignored whenever an animation to that object was done. That is now fixed. Also added the Selection object, to store selected objects, and the fromJSON() method to map objects, allowing you to input json into the object and set it's properties. Maybe some other random things too idk